Improved Data Structure Tests, found bug, fixed it
There now is a shared class DataValidator to be used for checking the
integrity of the experiment data structures. It is used in the JUnit
tests. This led to finding a bug in DataSelection, the creator of the
shadow data structures: parameter settings were sometimes lost. This bug
is fixed now.