Added Draft of Evaluation Modules and Jobs Interfaces and Base Classes
The first draft of the specification of the evaluation modules has been
added. This needs to be in the evaluator-base so that we can out source
the modules into a separate project and then have an overall
evaluator-main project using evaluator-base, evaluator-attributes, and
evaluator-modules.

I want to simplify the current definition from the main project
to make the system more versatile.