Improved Random Example
We now try to ensure that the experiments share runs for at least one
benchmark instance. Since they would be incomparable otherwise, this
should be a good idea. Also, if the instances for which experimental
runs exist are entirely disjoint, a couple of tests based on the data
might fail: We can, for example, not perform any behavioral clustering
in that case.