New Sanity for RandomExample Now Optional
The RandomExample for experiment data has been "sanitized" in some of
the previous commits. Proper testing of the experiment attribute
computations may require even more sanitization. However, testing IO,
for instance, does not. Matter of fact, IO should be fired at with data
as odd as possible. Therefore, I added a new parameter to the random
example which is used to decide how "random" the data will actually be.
This will allow me to add even more sanity restrictions in the future,
while still testing IO as hard as possible.